pos_backend/pos/requirements.txt
django
Pillow
djangorestframework
djangorestframework_simplejwt
django-cors-headers